+++ src/cmd/list.rs
@@ -1,6 +1,7 @@
 use anyhow::{bail, Context, Result};
 use clap::Clap;
 use console::Style;
+use regex::Regex;
 use ssh2::FileStat;
 use std::path::PathBuf;
 
@@ -13,7 +14,9 @@ use crate::ssh::{FileListing, SshSession};
 /// List uploaded files and their URLs.
 #[derive(Clap, Debug)]
 pub struct List {
-    /// Only list last `n` entries
+    /// Only list newest `n` entries. Note that entries are selected prior to sorting. That means
+    /// that if you want to get the largest files by size you should not specify `--last`.
+    /// Otherwise, only the last `<n>` files will be sorted by size.
     #[clap(short = 'n', long)]
     last: Option<usize>,
 
@@ -24,6 +27,18 @@ pub struct List {
     /// Show no full urls but rather filenames only. Makes for more concise output.
     #[clap(long, short)]
     filenames: bool,
+
+    /// Filter filenames by regex. See https://docs.rs/regex/latest/regex/#syntax
+    #[clap(long, short = 'F', value_name = "regex")]
+    filter: Option<String>,
+
+    /// Reverse listing.
+    #[clap(long, short)]
+    reverse: bool,
+
+    /// Sort listing by size
+    #[clap(long, short = 'S')]
+    sort_size: bool,
 
     /// Only list the remote URLs (useful for copying and scripting).
     #[clap(short, long = "url-only")]
@@ -38,7 +53,7 @@ impl Command for List {
     fn run(&self, session: &SshSession, _config: &Config) -> Result<()> {
         let host = &session.host;
 
-        let to_list: FileListing = if self.indices.len() == 0 {
+        let mut to_list: FileListing = if self.indices.len() == 0 {
             let files = session.list_files()?;
             let num_files = files.len();
             if let Some(n) = self.last {
@@ -60,6 +75,17 @@ impl Command for List {
             session.get_files_by(&self.indices, &[], session.host.prefix_length)?
         };
 
+        if self.filter.is_some() {
+            let re = Regex::new(&self.filter.as_ref().unwrap())?;
+            to_list.files = to_list
+                .files
+                .into_iter()
+                .filter(|(_, path)| {
+                    re.is_match(&path.file_name().unwrap().to_string_lossy().to_string())
+                })
+                .collect()
+        }
+
         let num_digits = {
             let mut num_digits = 0;
             let mut num = to_list.num_files;
@@ -75,7 +101,7 @@ impl Command for List {
                 println!("{}", host.get_url(&format!("{}", file.display()))?);
             }
         } else {
-            let list_infos: Vec<(&(usize, PathBuf), Option<ssh2::FileStat>)> = {
+            let mut list_infos: Vec<(&(usize, PathBuf), Option<ssh2::FileStat>)> = {
                 if self.stats_needed() {
                     let files = to_list.files.iter().map(|f| f.1.as_ref());
                     to_list
@@ -87,6 +113,15 @@ impl Command for List {
                     to_list.files.iter().zip(std::iter::repeat(None)).collect()
                 }
             };
+
+            if self.sort_size {
+                list_infos.sort_by_key(|(_, fs)| fs.as_ref().unwrap().size.unwrap());
+            }
+
+            if self.reverse {
+                list_infos.reverse();
+            }
+
             let content: Result<Vec<String>> = list_infos
                 .iter()
                 .map(|((i, file), stat)| -> Result<String> {
@@ -128,7 +163,7 @@ impl Command for List {
 impl List {
     /// Return whether or not we need to fetch stats
     fn stats_needed(&self) -> bool {
-        self.with_size
+        self.with_size || self.sort_size
     }
 
     fn size_column(&self, stat: &FileStat) -> Result<String> {